refactor(FollowUseCase): refactor ConcurrentAssociationUserRepository to FollowUseCase
