+++ app/src/main/java/com/android/unio/model/usecase/FollowUseCase.kt
@@ -1,4 +1,4 @@
-package com.android.unio.model.follow
+package com.android.unio.model.usecase
 
 import com.android.unio.model.association.Association
 import com.android.unio.model.user.User
