+++ app/src/test/java/com/android/unio/model/association/AssociationViewModelTest.kt
@@ -5,7 +5,7 @@ import com.android.unio.mocks.association.MockAssociation
 import com.android.unio.model.event.Event
 import com.android.unio.model.event.EventRepository
 import com.android.unio.model.firestore.emptyFirestoreReferenceList
-import com.android.unio.model.follow.ConcurrentAssociationUserRepository
+import com.android.unio.model.usecase.ConcurrentAssociationUserRepository
 import com.android.unio.model.image.ImageRepository
 import com.android.unio.model.user.User
 import com.google.firebase.Firebase
